+++ src/platform.ts
@@ -42,8 +42,6 @@ import {
   LevelControlCluster,
   BooleanStateCluster,
   ClusterRegistry,
-  SwitchCluster,
-  Switch,
   ColorControlCluster,
   electricalSensor,
   OccupancySensingCluster,
@@ -55,6 +53,9 @@ import {
   PowerSourceCluster,
   ElectricalPowerMeasurement,
   ElectricalEnergyMeasurement,
+  Endpoint,
+  ElectricalPowerMeasurementCluster,
+  ElectricalEnergyMeasurementCluster,
 } from 'matterbridge';
 
 import { EveHistory, EveHistoryCluster, MatterHistory } from 'matterbridge/history';
@@ -301,7 +302,7 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
         }
       }
 
-      // Create a new Matterbridge device for the switch
+      // Create a new Matterbridge device
       const mbDevice = new MatterbridgeDevice(bridgedNode, undefined, config.debug as boolean);
       mbDevice.createDefaultBridgedDeviceBasicInformationClusterServer(
         device.name,
@@ -370,23 +371,8 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
             const child = mbDevice.addChildDeviceTypeWithClusterServer(key, [deviceType], clusterIds);
             mbDevice.configureColorControlCluster(true, false, false, ColorControl.ColorMode.CurrentHueAndCurrentSaturation, child);
 
-            // Add the electrical EveHistory cluster
-            if (
-              config.exposePowerMeter === 'evehistory' &&
-              lightComponent.hasProperty('voltage') &&
-              lightComponent.hasProperty('current') &&
-              lightComponent.hasProperty('apower') &&
-              lightComponent.hasProperty('aenergy')
-            ) {
-              child.addClusterServer(
-                MatterHistory.getEveHistoryClusterServer(
-                  lightComponent.getValue('voltage') as number,
-                  lightComponent.getValue('current') as number,
-                  lightComponent.getValue('apower') as number,
-                  ((lightComponent.getValue('aenergy') as ShellyData).total as number) / 1000,
-                ),
-              );
-            }
+            // Add the electrical measurementa cluster on the same endpoint
+            this.addElectricalMeasurements(mbDevice, child, device, lightComponent);
 
             // Set the onOff attribute
             const state = lightComponent.getValue('state');
@@ -398,8 +384,6 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
               const matterLevel = Math.max(Math.min(Math.round(level / 100) * 255, 255), 0);
               child.getClusterServer(LevelControlCluster)?.setCurrentLevelAttribute(matterLevel);
             }
-            // TODO Set the currentHue and currentSaturation attribute
-            // TODO Set the currentX and currentY attribute
 
             // Add command handlers from Matter
             mbDevice.addCommandHandler('on', async (data) => {
@@ -464,23 +448,8 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
             if (config.exposeSwitch === 'outlet') deviceType = DeviceTypes.ON_OFF_PLUGIN_UNIT;
             const child = mbDevice.addChildDeviceTypeWithClusterServer(key, [deviceType], [OnOff.Cluster.id]);
 
-            // Add the electrical EveHistory cluster
-            if (
-              config.exposePowerMeter === 'evehistory' &&
-              switchComponent.hasProperty('voltage') &&
-              switchComponent.hasProperty('current') &&
-              switchComponent.hasProperty('apower') &&
-              switchComponent.hasProperty('aenergy')
-            ) {
-              child.addClusterServer(
-                MatterHistory.getEveHistoryClusterServer(
-                  switchComponent.getValue('voltage') as number,
-                  switchComponent.getValue('current') as number,
-                  switchComponent.getValue('apower') as number,
-                  ((switchComponent.getValue('aenergy') as ShellyData).total as number) / 1000,
-                ),
-              );
-            }
+            // Add the electrical measurementa cluster on the same endpoint
+            this.addElectricalMeasurements(mbDevice, child, device, switchComponent);
 
             // Set the OnOff attribute
             const state = switchComponent.getValue('state');
@@ -504,23 +473,8 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
           if (coverComponent) {
             const child = mbDevice.addChildDeviceTypeWithClusterServer(key, [DeviceTypes.WINDOW_COVERING], [WindowCovering.Cluster.id]);
 
-            // Add the electrical EveHistory cluster
-            if (
-              config.exposePowerMeter === 'evehistory' &&
-              coverComponent.hasProperty('voltage') &&
-              coverComponent.hasProperty('current') &&
-              coverComponent.hasProperty('apower') &&
-              coverComponent.hasProperty('aenergy')
-            ) {
-              child.addClusterServer(
-                MatterHistory.getEveHistoryClusterServer(
-                  coverComponent.getValue('voltage') as number,
-                  coverComponent.getValue('current') as number,
-                  coverComponent.getValue('apower') as number,
-                  ((coverComponent.getValue('aenergy') as ShellyData).total as number) / 1000,
-                ),
-              );
-            }
+            // Add the electrical measurementa cluster on the same endpoint
+            this.addElectricalMeasurements(mbDevice, child, device, coverComponent);
 
             // Set the WindowCovering attributes
             /*
@@ -562,105 +516,112 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
               // Add the Matter 1.3 electricalSensor device type with the ElectricalPowerMeasurement and ElectricalEnergyMeasurement clusters
               const child = mbDevice.addChildDeviceTypeWithClusterServer(key, [electricalSensor], [ElectricalPowerMeasurement.Cluster.id, ElectricalEnergyMeasurement.Cluster.id]);
 
-              // Set the electrical attributes
-              const epm = child.getClusterServer(ElectricalPowerMeasurement.Complete);
-              const voltage = pmComponent.hasProperty('voltage') ? pmComponent.getValue('voltage') : undefined;
-              if (isValidNumber(voltage, 0)) epm?.setVoltageAttribute(voltage * 1000);
+              device.log.info(
+                `***Added ElectricalPowerMeasurement and ElectricalEnergyMeasurement clusters to endpoint ${hk}${child.name}${nf} component ${hk}${component.name}:${component.id}${nf}`,
+              );
 
-              const current = pmComponent.hasProperty('current') ? pmComponent.getValue('current') : undefined;
-              if (isValidNumber(current, 0)) epm?.setActiveCurrentAttribute(current * 1000);
-
-              const power1 = pmComponent.hasProperty('power') ? pmComponent.getValue('power') : undefined; // Gen 1 devices
-              if (isValidNumber(power1, 0)) epm?.setActivePowerAttribute(power1 * 1000);
-              const power2 = pmComponent.hasProperty('apower') ? pmComponent.getValue('apower') : undefined; // Gen 2 devices
-              if (isValidNumber(power2, 0)) epm?.setActivePowerAttribute(power2 * 1000);
-
-              const eem = child.getClusterServer(ElectricalEnergyMeasurement.Complete);
-
-              const energy1 = pmComponent.hasProperty('total') ? pmComponent.getValue('total') : undefined; // Gen 1 devices in w/h
-              if (isValidNumber(energy1, 0)) eem?.setCumulativeEnergyImportedAttribute({ energy: energy1 * 1000 });
-
-              const energy2 = pmComponent.hasProperty('aenergy') ? pmComponent.getValue('aenergy') : undefined; // Gen 2 devices in w/h
-              if (isValidObject(energy2) && isValidNumber((energy2 as ShellyData).total, 0))
-                eem?.setCumulativeEnergyImportedAttribute({ energy: ((energy2 as ShellyData).total as number) * 1000 });
+              // Update the electrical attributes
+              for (const property of component.properties) {
+                if (!['voltage', 'current', 'power', 'apower', 'act_power', 'total', 'aenergy'].includes(property.key)) continue;
+                /*
+                device.log.info(
+                  `***Property ${property.key} value ${property.value !== null && typeof property.value === 'object' ? debugStringify(property.value as object) : property.value}${rs}`,
+                );
+                */
+                this.shellyUpdateHandler(mbDevice, device, component.id, property.key, property.value);
+              }
             } else if (config.exposePowerMeter === 'evehistory') {
               // Add the powerSource device type with the EveHistory cluster for HA
               ClusterRegistry.register(EveHistory.Complete);
               const child = mbDevice.addChildDeviceTypeWithClusterServer(key, [powerSource], [EveHistory.Cluster.id]);
-              // Set the electrical attributes
-              const voltage = pmComponent.hasProperty('voltage') ? pmComponent.getValue('voltage') : undefined;
-              if (isValidNumber(voltage, 0)) child.getClusterServer(EveHistoryCluster.with(EveHistory.Feature.EveEnergy))?.setVoltageAttribute(voltage);
+              device.log.info(`***Added EveHistory cluster to endpoint ${hk}${child.name}${nf} component ${hk}${component.name}:${component.id}${nf}`);
 
-              const current = pmComponent.hasProperty('current') ? pmComponent.getValue('current') : undefined;
-              if (isValidNumber(current, 0)) child.getClusterServer(EveHistoryCluster.with(EveHistory.Feature.EveEnergy))?.setCurrentAttribute(current);
-
-              const power1 = pmComponent.hasProperty('power') ? pmComponent.getValue('power') : undefined; // Gen 1 devices
-              if (isValidNumber(power1, 0)) child.getClusterServer(EveHistoryCluster.with(EveHistory.Feature.EveEnergy))?.setConsumptionAttribute(power1);
-              const power2 = pmComponent.hasProperty('apower') ? pmComponent.getValue('apower') : undefined; // Gen 2 devices
-              if (isValidNumber(power2, 0)) child.getClusterServer(EveHistoryCluster.with(EveHistory.Feature.EveEnergy))?.setConsumptionAttribute(power2);
-
-              const energy1 = pmComponent.hasProperty('total') ? pmComponent.getValue('total') : undefined; // Gen 1 devices in w/h
-              if (isValidNumber(energy1, 0)) child.getClusterServer(EveHistoryCluster.with(EveHistory.Feature.EveEnergy))?.setTotalConsumptionAttribute(energy1 / 1000);
-              const energy2 = pmComponent.hasProperty('aenergy') ? pmComponent.getValue('aenergy') : undefined; // Gen 2 devices in w/h
-              if (isValidObject(energy2) && isValidNumber((energy2 as ShellyData).total, 0))
-                child.getClusterServer(EveHistoryCluster.with(EveHistory.Feature.EveEnergy))?.setTotalConsumptionAttribute(((energy2 as ShellyData).total as number) / 1000);
+              // Update the electrical attributes
+              for (const property of component.properties) {
+                if (!['voltage', 'current', 'power', 'apower', 'act_power', 'total', 'aenergy'].includes(property.key)) continue;
+                /*
+                device.log.info(
+                  `***Property ${property.key} value ${property.value !== null && typeof property.value === 'object' ? debugStringify(property.value as object) : property.value}${rs}`,
+                );
+                */
+                this.shellyUpdateHandler(mbDevice, device, component.id, property.key, property.value);
+              }
             }
             // Add event handler
             pmComponent.on('update', (component: string, property: string, value: ShellyDataType) => {
               this.shellyUpdateHandler(mbDevice, device, component, property, value);
             });
           }
-        } else if (component.name === 'Input' && config.exposeInput !== 'disabled') {
+        } else if (component.name === 'Input') {
           const inputComponent = device.getComponent(key);
+          if (inputComponent && inputComponent?.hasProperty('enable') && inputComponent?.getValue('enable') === false) continue;
           if (inputComponent && inputComponent?.hasProperty('state') && config.exposeInput === 'contact') {
-            const child = mbDevice.addChildDeviceTypeWithClusterServer(key, [DeviceTypes.CONTACT_SENSOR], []);
-            // Set the state attribute
             const state = inputComponent.getValue('state') as boolean;
-            if (isValidBoolean(state)) child.getClusterServer(BooleanStateCluster)?.setStateValueAttribute(state);
-            // Add event handler
-            inputComponent.on('update', (component: string, property: string, value: ShellyDataType) => {
-              this.shellyUpdateHandler(mbDevice, device, component, property, value);
-            });
+            if (isValidBoolean(state)) {
+              const child = mbDevice.addChildDeviceTypeWithClusterServer(key, [DeviceTypes.CONTACT_SENSOR], []);
+              // Set the state attribute
+              child.getClusterServer(BooleanStateCluster)?.setStateValueAttribute(state);
+              // Add event handler
+              inputComponent.on('update', (component: string, property: string, value: ShellyDataType) => {
+                this.shellyUpdateHandler(mbDevice, device, component, property, value);
+              });
+            }
           } else if (inputComponent && inputComponent?.hasProperty('state') && config.exposeInput === 'momentary') {
-            const child = mbDevice.addChildDeviceTypeWithClusterServer(key, [DeviceTypes.GENERIC_SWITCH], []);
-            child.addClusterServer(mbDevice.getDefaultSwitchClusterServer());
-            // Set the state attribute
             const state = inputComponent.getValue('state') as boolean;
-            if (isValidBoolean(state)) child.getClusterServer(SwitchCluster.with(Switch.Feature.MomentarySwitch))?.setCurrentPositionAttribute(state ? 1 : 0);
-            // Add event handler
-            inputComponent.on('update', (component: string, property: string, value: ShellyDataType) => {
-              this.shellyUpdateHandler(mbDevice, device, component, property, value);
-            });
+            if (isValidBoolean(state)) {
+              const child = mbDevice.addChildDeviceTypeWithClusterServer(key, [DeviceTypes.GENERIC_SWITCH], []);
+              child.addClusterServer(mbDevice.getDefaultSwitchClusterServer());
+              // Add event handler
+              inputComponent.on('update', (component: string, property: string, value: ShellyDataType) => {
+                this.shellyUpdateHandler(mbDevice, device, component, property, value);
+              });
+            }
           } else if (inputComponent && inputComponent?.hasProperty('state') && config.exposeInput === 'latching') {
-            const child = mbDevice.addChildDeviceTypeWithClusterServer(key, [DeviceTypes.GENERIC_SWITCH], []);
-            child.addClusterServer(mbDevice.getDefaultLatchingSwitchClusterServer());
-            // Set the state attribute
             const state = inputComponent.getValue('state') as boolean;
-            if (isValidBoolean(state)) child.getClusterServer(SwitchCluster.with(Switch.Feature.LatchingSwitch))?.setCurrentPositionAttribute(state ? 1 : 0);
-            // Add event handler
-            inputComponent.on('update', (component: string, property: string, value: ShellyDataType) => {
-              this.shellyUpdateHandler(mbDevice, device, component, property, value);
-            });
+            if (isValidBoolean(state)) {
+              const child = mbDevice.addChildDeviceTypeWithClusterServer(key, [DeviceTypes.GENERIC_SWITCH], []);
+              child.addClusterServer(mbDevice.getDefaultLatchingSwitchClusterServer());
+              // Add event handler
+              inputComponent.on('update', (component: string, property: string, value: ShellyDataType) => {
+                this.shellyUpdateHandler(mbDevice, device, component, property, value);
+              });
+            }
           } else if (inputComponent && inputComponent?.hasProperty('event') && config.exposeInputEvent !== 'disabled') {
-            const child = mbDevice.addChildDeviceTypeWithClusterServer(key, [DeviceTypes.GENERIC_SWITCH], []);
-            child.addClusterServer(mbDevice.getDefaultSwitchClusterServer());
-            // Set the current position to 0
-            child.getClusterServer(SwitchCluster.with(Switch.Feature.MomentarySwitch))?.setCurrentPositionAttribute(0);
-            // Add event handler
-            inputComponent.on('update', (component: string, property: string, value: ShellyDataType) => {
-              this.shellyUpdateHandler(mbDevice, device, component, property, value);
-            });
+            // Gen 1 devices
+            const event = inputComponent.getValue('event') as boolean;
+            if (isValidString(event)) {
+              const child = mbDevice.addChildDeviceTypeWithClusterServer(key, [DeviceTypes.GENERIC_SWITCH], []);
+              child.addClusterServer(mbDevice.getDefaultSwitchClusterServer());
+              // Add event handler
+              inputComponent.on('update', (component: string, property: string, value: ShellyDataType) => {
+                this.shellyUpdateHandler(mbDevice, device, component, property, value);
+              });
+            }
           }
-        } else if (component.name === 'Input' && config.exposeInput === 'disabled') {
-          const inputComponent = device.getComponent(key);
-          if (inputComponent && inputComponent?.hasProperty('event') && config.exposeInputEvent !== 'disabled') {
+          if (
+            component &&
+            component.hasProperty('state') &&
+            component.getValue('state') === null &&
+            component.hasProperty('type') &&
+            component.getValue('type') === 'button' &&
+            config.exposeInputEvent !== 'disabled'
+          ) {
+            // Gen 2/3 devices with Input type=button
             const child = mbDevice.addChildDeviceTypeWithClusterServer(key, [DeviceTypes.GENERIC_SWITCH], []);
             child.addClusterServer(mbDevice.getDefaultSwitchClusterServer());
-            // Set the current position to 0
-            child.getClusterServer(SwitchCluster.with(Switch.Feature.MomentarySwitch))?.setCurrentPositionAttribute(0);
-            // Add event handler
-            inputComponent.on('update', (component: string, property: string, value: ShellyDataType) => {
-              this.shellyUpdateHandler(mbDevice, device, component, property, value);
+            device.log.info(`Add device event handler for device ${idn}${device.id}${rs} component ${hk}${component.id}${db} type Button`);
+            component.on('event', (component: string, event: string) => {
+              if (isValidString(component, 7) && isValidString(event, 9, 11) && device.getComponent(component)) {
+                device.log.info(`${db}Shelly event ${hk}${component}${db}:${zb}${event}${db} for device ${idn}${device.id}${rs}${db}`);
+                const endpoint = mbDevice.getChildEndpointByName(component);
+                if (!endpoint) {
+                  device.log.error(`getChildEndpointByName(${component}) for device ${idn}${device.id}${rs} failed`);
+                  return;
+                }
+                if (event === 'single_push') mbDevice.triggerSwitchEvent('Single', device.log, endpoint);
+                if (event === 'double_push') mbDevice.triggerSwitchEvent('Double', device.log, endpoint);
+                if (event === 'long_push') mbDevice.triggerSwitchEvent('Long', device.log, endpoint);
+              }
             });
           }
         } else if (component.name === 'Sensor' && config.exposeSensor !== 'disabled') {
@@ -847,20 +808,54 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
       if (!shellyDevice) return;
       mbDevice.getChildEndpoints().forEach(async (childEndpoint) => {
         const label = mbDevice.getEndpointLabel(childEndpoint.number);
+        // Configure the cluster OnOff attribute onOff
         if (label?.startsWith('switch') || label?.startsWith('relay') || label?.startsWith('light') || label?.startsWith('rgb')) {
           const switchComponent = shellyDevice.getComponent(label) as ShellySwitchComponent;
           this.log.info(`Configuring device ${dn}${mbDevice.deviceName}${nf} component ${hk}${label}${nf}:${zb}state ${YELLOW}${switchComponent.getValue('state')}${nf}`);
-          if (isValidBoolean(switchComponent.getValue('state'))) childEndpoint.getClusterServer(OnOffCluster)?.setOnOffAttribute(switchComponent.getValue('state') as boolean);
+          const state = switchComponent.getValue('state');
+          if (isValidBoolean(state)) {
+            mbDevice.setAttribute(OnOffCluster.id, 'onOff', state, shellyDevice.log, childEndpoint);
+          }
         }
+        // Configure the cluster LevelControl attribute currentLevel
         if (label?.startsWith('light') || label?.startsWith('rgb')) {
           const lightComponent = shellyDevice.getComponent(label) as ShellyLightComponent;
           const level = lightComponent.getValue('brightness') as number;
           if (isValidNumber(level, 0, 100)) {
             const matterLevel = Math.max(Math.min(Math.round((level / 100) * 255), 255), 0);
             this.log.info(`Configuring device ${dn}${mbDevice.deviceName}${nf} component ${hk}${label}${nf}:${zb}brightness ${YELLOW}${matterLevel}${nf}`);
-            childEndpoint.getClusterServer(LevelControlCluster)?.setCurrentLevelAttribute(matterLevel);
+            mbDevice.setAttribute(LevelControlCluster.id, 'currentLevel', matterLevel, shellyDevice.log, childEndpoint);
+          }
+          if (lightComponent.hasProperty('red') && lightComponent.hasProperty('green') && lightComponent.hasProperty('blue') && shellyDevice.profile !== 'white') {
+            const red = lightComponent.getValue('red') as number;
+            const green = lightComponent.getValue('green') as number;
+            const blue = lightComponent.getValue('blue') as number;
+            if (isValidNumber(red, 0, 255) && isValidNumber(green, 0, 255) && isValidNumber(blue, 0, 255)) {
+              this.log.info(`Configuring device ${dn}${mbDevice.deviceName}${nf} component ${hk}${label}${nf}:${zb}rgb ${YELLOW}${red},${green},${blue}${nf}`);
+              const hsl = rgbColorToHslColor({ r: red, g: green, b: blue });
+              this.log.debug(`ColorRgbToHsl: R:${red} G:${green} B:${blue} => H:${hsl.h} S:${hsl.s} L:${hsl.l}`);
+              const hue = Math.max(Math.min(Math.round((hsl.h / 360) * 254), 254), 0);
+              const saturation = Math.max(Math.min(Math.round((hsl.s / 100) * 254), 254), 0);
+              if (isValidNumber(hue, 0, 254)) mbDevice.setAttribute(ColorControlCluster.id, 'currentHue', hue, shellyDevice.log, childEndpoint);
+              if (isValidNumber(saturation, 0, 254)) mbDevice.setAttribute(ColorControlCluster.id, 'currentSaturation', saturation, shellyDevice.log, childEndpoint);
+              mbDevice.setAttribute(ColorControlCluster.id, 'colorMode', ColorControl.ColorMode.CurrentHueAndCurrentSaturation, shellyDevice.log, childEndpoint);
+            }
+          }
+          if (lightComponent.hasProperty('rgb') && shellyDevice.profile !== 'white') {
+            const rgb = lightComponent.getValue('rgb') as object;
+            if (isValidArray(rgb, 3, 3) && isValidNumber(rgb[0], 0, 255) && isValidNumber(rgb[1], 0, 255) && isValidNumber(rgb[2], 0, 255)) {
+              this.log.info(`Configuring device ${dn}${mbDevice.deviceName}${nf} component ${hk}${label}${nf}:${zb}rgb ${YELLOW}${rgb[0]},${rgb[1]},${rgb[2]}${nf}`);
+              const hsl = rgbColorToHslColor({ r: rgb[0], g: rgb[1], b: rgb[2] });
+              this.log.debug(`ColorRgbToHsl: R:${rgb[0]} G:${rgb[1]} B:${rgb[2]} => H:${hsl.h} S:${hsl.s} L:${hsl.l}`);
+              const hue = Math.max(Math.min(Math.round((hsl.h / 360) * 254), 254), 0);
+              const saturation = Math.max(Math.min(Math.round((hsl.s / 100) * 254), 254), 0);
+              if (isValidNumber(hue, 0, 254)) mbDevice.setAttribute(ColorControlCluster.id, 'currentHue', hue, shellyDevice.log, childEndpoint);
+              if (isValidNumber(saturation, 0, 254)) mbDevice.setAttribute(ColorControlCluster.id, 'currentSaturation', saturation, shellyDevice.log, childEndpoint);
+              mbDevice.setAttribute(ColorControlCluster.id, 'colorMode', ColorControl.ColorMode.CurrentHueAndCurrentSaturation, shellyDevice.log, childEndpoint);
+            }
           }
         }
+        // Configure the cluster WindowCovering attribute currentPositionLiftPercent100ths
         if (label?.startsWith('cover') || label?.startsWith('roller')) {
           const coverComponent = shellyDevice.getComponent(label) as ShellyCoverComponent;
           const position = coverComponent.hasProperty('current_pos') ? (coverComponent.getValue('current_pos') as number) : undefined;
@@ -872,6 +867,7 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
             mbDevice.setWindowCoveringTargetAsCurrentAndStopped(childEndpoint);
           }
         }
+        /* No need for configuration, the cluster already has the state
         if (label?.startsWith('input')) {
           const inputComponent = shellyDevice.getComponent(label) as ShellyComponent;
           if (inputComponent.hasProperty('state') && isValidBoolean(inputComponent.getValue('state'))) {
@@ -881,6 +877,7 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
               childEndpoint.getClusterServer(Switch.Complete)?.setCurrentPositionAttribute(inputComponent.getValue('state') === true ? 1 : 0);
           }
         }
+        */
       });
     });
   }
@@ -900,7 +897,7 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
     this.shelly.setLogLevel(logLevel, this.config.debugMdns as boolean, this.config.debugCoap as boolean, this.config.debugWs as boolean);
   }
 
-  localVerifyMatterbridgeVersion(version: string): boolean {
+  localVerifyMatterbridgeVersion(requiredVersion: string): boolean {
     const compareVersions = (matterbridgeVersion: string, requiredVersion: string): boolean => {
       const stripTag = (v: string) => {
         const parts = v.split('-');
@@ -920,8 +917,47 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
       return true;
     };
 
-    if (!compareVersions(this.matterbridge.matterbridgeVersion, version)) return false;
+    if (!compareVersions(this.matterbridge.matterbridgeVersion, requiredVersion)) return false;
     return true;
+  }
+
+  private addElectricalMeasurements(device: MatterbridgeDevice, endpoint: Endpoint | undefined, shelly: ShellyDevice, component: ShellyComponent) {
+    if (!endpoint) {
+      // this.log.info(`addElectricalMeasurements: endpoint is undefined`);
+      return;
+    }
+    const updateProperties = () => {
+      for (const property of component.properties) {
+        if (!['voltage', 'current', 'power', 'apower', 'act_power', 'total', 'aenergy'].includes(property.key)) continue;
+        /*
+        shelly.log.info(
+          `***Property ${property.key} value ${property.value !== null && typeof property.value === 'object' ? debugStringify(property.value as object) : property.value}${rs}`,
+        );
+        */
+        this.shellyUpdateHandler(device, shelly, component.id, property.key, property.value);
+      }
+    };
+
+    // Add the electrical EveHistory cluster on the same endpoint
+    if (
+      this.config.exposePowerMeter === 'evehistory' &&
+      (component.hasProperty('voltage') || component.hasProperty('current') || component.hasProperty('apower') || component.hasProperty('aenergy'))
+    ) {
+      shelly.log.info(`***Adding EveHistory cluster to endpoint ${hk}${endpoint.name}${nf} component ${hk}${component.id}${nf}`);
+      endpoint.addClusterServer(MatterHistory.getEveHistoryClusterServer());
+      updateProperties();
+    }
+
+    // Add the Matter 1.3 ElectricalPowerMeasurement and ElectricalEnergyMeasurement cluster on the same endpoint
+    if (
+      this.config.exposePowerMeter === 'matter13' &&
+      (component.hasProperty('voltage') || component.hasProperty('current') || component.hasProperty('apower') || component.hasProperty('aenergy'))
+    ) {
+      shelly.log.info(`***Adding ElectricalPowerMeasurement and ElectricalEnergyMeasurement clusters to endpoint ${hk}${endpoint.name}${nf} component ${hk}${component.id}${nf}`);
+      endpoint.addClusterServer(device.getDefaultElectricalPowerMeasurementClusterServer());
+      endpoint.addClusterServer(device.getDefaultElectricalEnergyMeasurementClusterServer());
+      updateProperties();
+    }
   }
 
   private async saveStoredDevices(): Promise<boolean> {
@@ -1113,7 +1149,7 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
         const hue = Math.max(Math.min(Math.round((hsl.h / 360) * 254), 254), 0);
         const saturation = Math.max(Math.min(Math.round((hsl.s / 100) * 254), 254), 0);
         if (isValidNumber(hue, 0, 254)) matterbridgeDevice.setAttribute(ColorControlCluster.id, 'currentHue', hue, shellyDevice.log, endpoint);
-        if (isValidNumber(hue, 0, 254)) matterbridgeDevice.setAttribute(ColorControlCluster.id, 'currentSaturation', saturation, shellyDevice.log, endpoint);
+        if (isValidNumber(saturation, 0, 254)) matterbridgeDevice.setAttribute(ColorControlCluster.id, 'currentSaturation', saturation, shellyDevice.log, endpoint);
         matterbridgeDevice.setAttribute(ColorControlCluster.id, 'colorMode', ColorControl.ColorMode.CurrentHueAndCurrentSaturation, shellyDevice.log, endpoint);
       }, 200);
     }
@@ -1281,9 +1317,14 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
       */
     }
     // Update energy from main components (gen 2 devices send power total inside the component not with meter)
-    if (this.config.exposePowerMeter === 'evehistory' && ['Light', 'Relay', 'Switch', 'Cover', 'Roller', 'PowerMeter'].includes(shellyComponent.name)) {
+    if (
+      (this.config.exposePowerMeter === 'evehistory' || this.config.exposePowerMeter === 'matter13') &&
+      ['Light', 'Rgb', 'Relay', 'Switch', 'Cover', 'Roller', 'PowerMeter'].includes(shellyComponent.name)
+    ) {
       if ((property === 'power' || property === 'apower' || property === 'act_power') && isValidNumber(value, 0)) {
-        matterbridgeDevice.setAttribute(EveHistoryCluster.id, 'consumption', value, shellyDevice.log, endpoint);
+        if (this.config.exposePowerMeter === 'evehistory') matterbridgeDevice.setAttribute(EveHistoryCluster.id, 'consumption', value, shellyDevice.log, endpoint);
+        else if (this.config.exposePowerMeter === 'matter13')
+          matterbridgeDevice.setAttribute(ElectricalPowerMeasurementCluster.id, 'activePower', value * 1000, shellyDevice.log, endpoint);
         if (property === 'act_power') return; // Skip the rest for PRO devices
         if (shellyComponent.id.startsWith('emeter')) return; // Skip the rest for em3 devices
         // Calculate current from power and voltage
@@ -1291,32 +1332,44 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
         if (isValidNumber(voltage, 10)) {
           let current = value / voltage;
           current = Math.round(current * 10000) / 10000; // round to 4 decimals
-          matterbridgeDevice.setAttribute(EveHistoryCluster.id, 'current', current, shellyDevice.log, endpoint);
+          if (this.config.exposePowerMeter === 'evehistory') matterbridgeDevice.setAttribute(EveHistoryCluster.id, 'current', current, shellyDevice.log, endpoint);
+          else if (this.config.exposePowerMeter === 'matter13')
+            matterbridgeDevice.setAttribute(ElectricalPowerMeasurementCluster.id, 'activeCurrent', current * 1000, shellyDevice.log, endpoint);
         }
       }
       if (property === 'total' && isValidNumber(value, 0)) {
         let energy = value / 1000; // convert to kWh
         energy = Math.round(energy * 10000) / 10000; // round to 4 decimals
-        matterbridgeDevice.setAttribute(EveHistoryCluster.id, 'totalConsumption', energy, shellyDevice.log, endpoint);
+        if (this.config.exposePowerMeter === 'evehistory') matterbridgeDevice.setAttribute(EveHistoryCluster.id, 'totalConsumption', energy, shellyDevice.log, endpoint);
+        else if (this.config.exposePowerMeter === 'matter13')
+          matterbridgeDevice.setAttribute(ElectricalEnergyMeasurementCluster.id, 'cumulativeEnergyImported', { energy: energy * 1000 }, shellyDevice.log, endpoint);
       }
       if (property === 'aenergy' && isValidObject(value) && isValidNumber((value as ShellyData).total, 0)) {
         let energy = ((value as ShellyData).total as number) / 1000; // convert to kWh
         energy = Math.round(energy * 10000) / 10000;
-        matterbridgeDevice.setAttribute(EveHistoryCluster.id, 'totalConsumption', energy, shellyDevice.log, endpoint);
+        if (this.config.exposePowerMeter === 'evehistory') matterbridgeDevice.setAttribute(EveHistoryCluster.id, 'totalConsumption', energy, shellyDevice.log, endpoint);
+        else if (this.config.exposePowerMeter === 'matter13')
+          matterbridgeDevice.setAttribute(ElectricalEnergyMeasurementCluster.id, 'cumulativeEnergyImported', { energy: energy * 1000 }, shellyDevice.log, endpoint);
       }
-      if (property === 'voltage' && isValidNumber(value, 10)) {
-        matterbridgeDevice.setAttribute(EveHistoryCluster.id, 'voltage', value, shellyDevice.log, endpoint);
+      if (property === 'voltage' && isValidNumber(value, 0)) {
+        if (this.config.exposePowerMeter === 'evehistory') matterbridgeDevice.setAttribute(EveHistoryCluster.id, 'voltage', value, shellyDevice.log, endpoint);
+        else if (this.config.exposePowerMeter === 'matter13')
+          matterbridgeDevice.setAttribute(ElectricalPowerMeasurementCluster.id, 'voltage', value * 1000, shellyDevice.log, endpoint);
       }
       if (property === 'current' && isValidNumber(value, 0)) {
-        matterbridgeDevice.setAttribute(EveHistoryCluster.id, 'current', value, shellyDevice.log, endpoint);
+        if (this.config.exposePowerMeter === 'evehistory') matterbridgeDevice.setAttribute(EveHistoryCluster.id, 'current', value, shellyDevice.log, endpoint);
+        else if (this.config.exposePowerMeter === 'matter13')
+          matterbridgeDevice.setAttribute(ElectricalPowerMeasurementCluster.id, 'activeCurrent', value * 1000, shellyDevice.log, endpoint);
         if (shellyComponent.hasProperty('act_power')) return; // Skip the rest for PRO devices
         if (shellyComponent.id.startsWith('emeter')) return; // Skip the rest for em3 devices
         // Calculate power from current and voltage
         const voltage = shellyComponent.hasProperty('voltage') ? (shellyComponent.getValue('voltage') as number) : undefined;
-        if (isValidNumber(voltage, 10)) {
+        if (isValidNumber(voltage, 0)) {
           let power = value * voltage;
           power = Math.round(power * 10000) / 10000;
-          matterbridgeDevice.setAttribute(EveHistoryCluster.id, 'consumption', power, shellyDevice.log, endpoint);
+          if (this.config.exposePowerMeter === 'evehistory') matterbridgeDevice.setAttribute(EveHistoryCluster.id, 'consumption', power, shellyDevice.log, endpoint);
+          else if (this.config.exposePowerMeter === 'matter13')
+            matterbridgeDevice.setAttribute(ElectricalPowerMeasurementCluster.id, 'activePower', power * 1000, shellyDevice.log, endpoint);
         }
       }
     }